api1/requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
openai
requests
colorama
docx